popImage 리덕스에 위젯 데이터 저장하는 부분 함수로 분리
