+++ src/components/Widgets/Image/PopImage.js
@@ -1,55 +1,32 @@
 /** @jsxImportSource @emotion/react */
 import React, { useEffect, useState } from 'react';
 import { css } from '@emotion/react';
-import { useDispatch, useSelector } from 'react-redux';
 import { InitButtonStyle } from '../../../styles/GlobalStyles';
-import { createReplacementWidgetsAction } from '../../../redux/slice';
-import {
-  ACTION_EDIT,
-  ACTION_NONE,
-  TYPE_IMAGE,
-} from '../../../utils/constantValue';
 import PopButtonsWrapper from '../PopButtonsWrapper';
-import { usePostImage } from '../../../hooks/widget';
+import { useEditWidget, usePostImage } from '../../../hooks/widget';
 
 function PopImage(props) {
-  const { widgets, modal } = useSelector((state) => ({
-    widgets: state.info.widgets,
-    modal: state.info.modal,
-  }));
-
+  const [submitted, setSubmitted] = useState(false);
   const [isLocalUpload, setIsLocalUpload] = useState(true);
   const [thumbnail, setThumbnail] = useState('');
+  const [localFiles, setLocalFiles] = useState(null);
   const { s3url, request } = usePostImage();
-  const dispatch = useDispatch();
 
-  function editWidget() {
-    const allWidgets = JSON.parse(JSON.stringify(widgets.list));
-    const targetId = modal.imgChangeTargetId;
-    const targetItem = allWidgets.find((widget) => widget.i === targetId);
-    targetItem.widget_type = TYPE_IMAGE;
-    targetItem.widget_data = {
-      thumbnail: `${thumbnail}`,
-      url: '',
-    };
-    if (
-      targetItem.widget_action === ACTION_NONE ||
-      targetItem.widget_code !== ''
-    ) {
-      targetItem.widget_action = ACTION_EDIT;
-    }
-    dispatch(
-      createReplacementWidgetsAction({
-        ...widgets,
-        list: allWidgets,
-      })
-    );
-  }
+  const { edit } = useEditWidget();
 
   const handleSubmit = () => {
-    editWidget();
-    props.endPop();
+    if (isLocalUpload && localFiles) {
+      request(localFiles); // post to server
+    }
+    setSubmitted(true);
   };
+
+  useEffect(() => {
+    if (submitted && thumbnail) {
+      edit(thumbnail);
+      props.endPop();
+    }
+  }, [thumbnail, submitted]);
 
   const handleThumbChange = ({ target: { value } }) => {
     setThumbnail(value);
@@ -62,7 +39,7 @@ function PopImage(props) {
   }, [s3url]);
 
   const handleLocalUpload = ({ target: { files } }) => {
-    request(files); // post to server
+    setLocalFiles(files);
   };
 
   const handleKeyDown = (event) => {
@@ -113,7 +90,6 @@ function PopImage(props) {
           type='button'
           onClick={() => {
             handleSubmit();
-            props.endPop();
           }}
         >
           확인
